# backend/docs/swagger.yaml
basePath: /
definitions:
  http.APIEndpointListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.APIEndpointResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.APIEndpointResp:
    properties:
      created_at:
        type: string
      datasource_id:
        type: integer
      description:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      methods:
        items:
          type: string
        type: array
      param_defs:
        items:
          $ref: '#/definitions/http.ParamDefResp'
        type: array
      params:
        items:
          type: string
        type: array
      path:
        type: string
      post_script_id:
        type: integer
      pre_script_id:
        type: integer
      project_id:
        type: integer
      sql:
        type: string
      summary:
        type: string
      tenant_id:
        type: integer
    type: object
  http.APIGroupListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.APIGroupResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.APIGroupResp:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      tenant_id:
        type: integer
    type: object
  http.AuthResp:
    properties:
      tenant:
        $ref: '#/definitions/http.TenantResp'
      tenants:
        items:
          $ref: '#/definitions/http.TenantResp'
        type: array
      token:
        type: string
      user:
        $ref: '#/definitions/http.UserResp'
    type: object
...
  http.PaginationInfo:
    properties:
      is_pager:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  http.ParamDefResp:
    properties:
      default:
        type: string
      desc:
        type: string
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
...
  http.R:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  http.RAPIEndpoint:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.APIEndpointResp'
      msg:
        type: string
    type: object
  http.RAPIEndpointList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.APIEndpointListResp'
      msg:
        type: string
    type: object
  http.RAPIGroup:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.APIGroupResp'
      msg:
        type: string
    type: object
  http.RAPIGroupList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.APIGroupListResp'
      msg:
        type: string
    type: object
  http.RAuth:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.AuthResp'
      msg:
        type: string
    type: object
...
  http.RTenant:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.TenantResp'
      msg:
        type: string
    type: object
...
  http.TenantResp:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  http.UserResp:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_superadmin:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: 企业级 SQL to API 智能网关平台管理接口
  title: OwlApi Control Plane
  version: "1.0"
paths:
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      operationId: login
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAuth'
      summary: 用户登录
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      operationId: register
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
            tenant_name:
              type: string
            tenant_slug:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAuth'
      summary: 注册用户
      tags:
      - auth
...
  /v1/tenants/{slug}:
    delete:
      operationId: deleteTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除租户 (SuperAdmin)
      tags:
      - tenant
    get:
      operationId: getTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 获取租户详情
      tags:
      - tenant
    put:
      consumes:
      - application/json
      operationId: updateTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
            plan:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 更新租户 (SuperAdmin)
      tags:
      - tenant
...
  /v1/tenants/{slug}/projects/{projectId}/endpoints:
    get:
      operationId: listEndpoints
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAPIEndpointList'
      security:
      - BearerAuth: []
      summary: 获取 API 端点列表
      tags:
      - endpoint
    post:
      consumes:
      - application/json
      operationId: createEndpoint
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 端点信息
        in: body
        name: body
        required: true
        schema:
          properties:
            datasource_id:
              type: integer
            description:
              type: string
            group_id:
              type: integer
            methods:
              type: array
            param_defs:
              type: array
            path:
              type: string
            post_script_id:
              type: integer
            pre_script_id:
              type: integer
            sql:
              type: string
            summary:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAPIEndpoint'
      security:
      - BearerAuth: []
      summary: 创建 API 端点
      tags:
      - endpoint
  /v1/tenants/{slug}/projects/{projectId}/endpoints/{endpointId}:
    delete:
      operationId: deleteEndpoint
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 端点ID
        in: path
        name: endpointId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除 API 端点
      tags:
      - endpoint
    put:
      consumes:
      - application/json
      operationId: updateEndpoint
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 端点ID
        in: path
        name: endpointId
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        schema:
          properties:
            datasource_id:
              type: integer
            description:
              type: string
            group_id:
              type: integer
            methods:
              type: array
            param_defs:
              type: array
            path:
              type: string
            post_script_id:
              type: integer
            pre_script_id:
              type: integer
            sql:
              type: string
            summary:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAPIEndpoint'
      security:
      - BearerAuth: []
      summary: 更新 API 端点
      tags:
      - endpoint
  /v1/tenants/{slug}/projects/{projectId}/groups:
    get:
      operationId: listGroups
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAPIGroupList'
      security:
      - BearerAuth: []
      summary: 获取 API 分组列表
      tags:
      - api-group
    post:
      consumes:
      - application/json
      operationId: createGroup
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 分组信息
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAPIGroup'
      security:
      - BearerAuth: []
      summary: 创建 API 分组
      tags:
      - api-group
  /v1/tenants/{slug}/projects/{projectId}/groups/{groupId}:
    delete:
      operationId: deleteGroup
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 分组ID
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除 API 分组
      tags:
      - api-group
    put:
      consumes:
      - application/json
      operationId: updateGroup
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 分组ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RAPIGroup'
      security:
      - BearerAuth: []
      summary: 更新 API 分组
      tags:
      - api-group
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"